// src/api/generated/models/authProfileResponseDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Unibac API
 * Autenticación: POST /auth/login fija la cookie HttpOnly access_token. En Swagger UI, ejecuta login en este mismo origen y luego las rutas protegidas enviarán la cookie automáticamente. Desde el frontend usa fetch/axios con credentials: include. Registro público: POST /auth/register-public solo si PUBLIC_REGISTRATION_ENABLED=true.
 * OpenAPI spec version: 1.0
 */
import type { AuthProfileResponseDtoCategoria } from './authProfileResponseDtoCategoria';
import type { AuthProfileResponseDtoNivel } from './authProfileResponseDtoNivel';
import type { AuthProfileResponseDtoRolId } from './authProfileResponseDtoRolId';
import type { AuthProfileResponseDtoTipo } from './authProfileResponseDtoTipo';

export interface AuthProfileResponseDto {
  id: number;
  usuario: string;
  nivel: AuthProfileResponseDtoNivel;
  tipo: AuthProfileResponseDtoTipo;
  rolId?: AuthProfileResponseDtoRolId;
  categoria?: AuthProfileResponseDtoCategoria;
}